Refactor and adds host dependencies.
