+++ .gitignore
@@ -2,3 +2,4 @@
 .cache
 *.sum
 *.retry
+*.vscode
