+++ group_vars/dev/main.yml
@@ -1,11 +1,1 @@
 ---
-Parameters:
-  EbAppName:
-    Type: String
-Type: AWS::EKS::Cluster
-Properties: 
-  Name: String
-  ResourcesVpcConfig: 
-    ResourcesVpcConfig
-  RoleArn: String
-  Version: String
