+++ roles/deployment/templates/deployment.yml.j2
@@ -1,0 +1,28 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: spleen
+  labels:
+    app: spleen
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: spleen
+  template:
+    metadata:
+      labels:
+        app: spleen
+    spec:
+      containers:
+      - name: spleen-app
+        image: {{ app_image }}
+        env:
+        - name: APP_PORT
+          value: {{ spleen_app_port }}
+        - name: APP_BUCKET
+          value: {{ spleen_app_bucket }}
+        - name: APP_KEY
+          value: {{ spleen_app_key }}
+        ports:
+        - containerPort: {{ spleen_app_port }}
