+++ roles/deployment/templates/service.yml.j2
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: spleen-service
+spec:
+  selector:
+    app: spleen
+  ports:
+  - protocol: TCP
+    port: {{ spleen_app_port }}
+    targetPort: {{ spleen_app_port }}
+  type: LoadBalancer
