+++ roles/deployment/vars/main.yml
@@ -1,0 +1,6 @@
+---
+app_tag: 1
+app_image: 409153575197.dkr.ecr.us-east-1.amazonaws.com/{{ ecr_name }}:{{ app_tag }}
+spleen_app_port: 8082
+spleen_app_bucket: "{{ spleen_bucket }}"
+spleen_app_key: "{{ spleen_key }}"
