+++ roles/docker/tasks/main.yml
@@ -1,10 +1,8 @@
 ---
-- name: Build Docker Image
-  include: build.yml
+- include: build.yml
   tags:
   - build   
 
-- name: Build Docker Image
-  include: run.yml
+- include: run.yml
   tags:
   - run
