--- roles/eks/templates/eks-nodegroup.yml
@@ -1,55 +1,0 @@
----
-AWSTemplateFormatVersion: "2010-09-09"
-Description: "Amazon EKS - Node Group"
-Parameters:
-  ClusterName:
-    Type: String
-  NodeAutoScalingGroupDesiredCapacity:
-    Type: Number
-  NodeAutoScalingGroupMaxSize:
-    Type: Number
-  NodeAutoScalingGroupMinSize:
-    Type: Number
-  NodeGroupName:
-    Type: String
-    Description: Unique identifier for the Node Group.
-  NodeInstanceType:
-    Type: String
-  NodeVolumeSize:
-    Type: Number
-  Subnet1:
-    Type: String
-  Subnet2:
-    Type: String
-  NodeRoleArn:
-    Type: String
-Resources:
-  NodeGroup:
-    Type: AWS::EKS::Nodegroup
-    Properties: 
-      AmiType: AL2_x86_64
-      ClusterName:
-        Fn::ImportValue: !Ref ClusterName
-      DiskSize: !Ref NodeVolumeSize
-      ForceUpdateEnabled: true
-      InstanceTypes: 
-        - !Ref NodeInstanceType
-      NodegroupName: !Ref NodeGroupName
-      NodeRole:
-        Fn::ImportValue: !Ref NodeRoleArn
-      ReleaseVersion: 1.14.7-20190927
-      ScalingConfig: 
-        MinSize: !Ref NodeAutoScalingGroupMinSize
-        DesiredSize: !Ref NodeAutoScalingGroupDesiredCapacity
-        MaxSize: !Ref NodeAutoScalingGroupMaxSize
-      Labels:
-        Key1: SpleenDemoTest
-      Subnets: 
-        - Fn::ImportValue: !Ref Subnet1
-        - Fn::ImportValue: !Ref Subnet2
-Outputs:
-  NodeGroup:
-    Description: EKS Node Group
-    Value: !Ref NodeGroup
-    Export:
-      Name: !Sub "${AWS::StackName}-${NodeGroupName}"
