+++ roles/eks/templates/eks_cluster_cf.yml
@@ -1,0 +1,42 @@
+---
+AWSTemplateFormatVersion: "2010-09-09"
+Description: "Amazon EKS Cluster"
+Parameters:
+  ClusterName:
+    Type: String
+  RoleArn:
+    Type: String
+  Version:
+    Type: String
+  PublicSubnet1:
+    Type: String
+  PublicSubnet2:
+    Type: String
+  PrivateSubnet1:
+    Type: String
+  PrivateSubnet2:
+    Type: String
+  EKSControlPlaneSG:
+    Type: String
+Resources:
+  EKSCluster:
+    Type: AWS::EKS::Cluster
+    Properties: 
+      Name: !Ref ClusterName
+      ResourcesVpcConfig:
+        SecurityGroupIds:
+          - Fn::ImportValue: !Ref EKSControlPlaneSG
+        SubnetIds:
+          - Fn::ImportValue: !Ref PublicSubnet1
+          - Fn::ImportValue: !Ref PrivateSubnet1
+          - Fn::ImportValue: !Ref PublicSubnet2
+          - Fn::ImportValue: !Ref PrivateSubnet2
+      RoleArn:
+        Fn::ImportValue: !Ref RoleArn
+      Version: !Ref Version
+Outputs:
+  EKSCluster:
+    Description: EKS Cluster
+    Value: !Ref EKSCluster
+    Export:
+      Name: !Sub "${AWS::StackName}-${ClusterName}"
