+++ roles/host_dependencies/tasks/main.yml
@@ -1,0 +1,4 @@
+---
+- include: golang.yml
+  
+- include: eksctl.yml
