--- roles/iam/tasks/roles.yml
@@ -1,11 +1,0 @@
----
-- name: IAM Roles Cloudformation Stack
-  cloudformation:
-    profile: "{{ profile }}"
-    region: "{{ region }}"
-    state: present
-    stack_name: "{{ roles_stack_name }}"
-    template: roles/iam/templates/roles.yml
-    template_parameters:          
-      EKSServiceRoleName: "{{ eks_service_role }}"
-      EKSWorkerNodeName: "{{ eks_worker_node_role }}"
