--- roles/iam/templates/roles.yml
@@ -1,51 +1,0 @@
----
-AWSTemplateFormatVersion: "2010-09-09"
-Description: "Amazon EKS Service Role"
-Parameters:
-  EKSServiceRoleName:
-    Type: String
-  EKSWorkerNodeName:
-    Type: String
-Resources:
-  eksServiceRole:
-    Type: AWS::IAM::Role
-    Properties:
-      AssumeRolePolicyDocument:
-        Version: "2012-10-17"
-        Statement:
-        - Effect: Allow
-          Principal:
-            Service:
-            - eks.amazonaws.com
-          Action:
-          - sts:AssumeRole
-      ManagedPolicyArns:
-        - arn:aws:iam::aws:policy/AmazonEKSServicePolicy
-        - arn:aws:iam::aws:policy/AmazonEKSClusterPolicy
-  eksWorkerNodeRole:
-    Type: AWS::IAM::Role
-    Properties:
-      AssumeRolePolicyDocument:
-        Version: "2012-10-17"
-        Statement:
-        - Effect: Allow
-          Principal:
-            Service:
-            - ec2.amazonaws.com
-          Action:
-          - sts:AssumeRole
-      ManagedPolicyArns:
-        - arn:aws:iam::aws:policy/AmazonEKSWorkerNodePolicy
-        - arn:aws:iam::aws:policy/AmazonEKS_CNI_Policy
-        - arn:aws:iam::aws:policy/AmazonEC2ContainerRegistryReadOnly
-Outputs:
-  RoleArn:
-    Description: The role that Amazon EKS will use to create AWS resources for Kubernetes clusters
-    Value: !GetAtt eksServiceRole.Arn
-    Export:
-      Name: !Sub "${AWS::StackName}-${EKSServiceRoleName}"
-  EKSWorkerNodeARN:
-    Description: The role that Amazon EKS will use to create AWS resources for Kubernetes clusters
-    Value: !GetAtt eksWorkerNodeRole.Arn
-    Export:
-      Name: !Sub "${AWS::StackName}-${EKSWorkerNodeName}"
