--- roles/network/tasks/main.yml
@@ -1,3 +1,0 @@
----
-- name: Create VPC
-  include: vpc.yml
