--- roles/network/tasks/vpc.yml
@@ -1,18 +1,0 @@
----
-- name: VPC Cloudformation Stack
-  cloudformation:
-    profile: "{{ profile }}"
-    region: "{{ region }}"
-    state: present
-    stack_name: "{{ vpc_stack_name }}"
-    template: roles/network/templates/vpc.yml
-    template_parameters:
-      EKSWorkerSG: "{{ eks_worker_sg }}"
-      EKSControlPlaneSG: "{{ eks_control_plane_sg }}"
-      PublicSubnet1Name: "{{ public_subnet_1 }}"
-      PrivateSubnet1Name: "{{ private_subnet_1 }}"
-      PublicSubnet2Name: "{{ public_subnet_2 }}"
-      PrivateSubnet2Name: "{{ private_subnet_2 }}"
-      Env: "{{ env | title }}"
-      Region: "{{ region }}"
-      EKSSubnetTag: "{{ eks_subnet_tag }}"
