+++ roles/s3/templates/s3_cf.yml
@@ -1,0 +1,19 @@
+---
+AWSTemplateFormatVersion: "2010-09-09"
+Description: Amazon S3 Buckets
+Parameters:
+  S3Bucket:
+    Type: String
+Resources:
+  SpleenBucket:
+    Type: AWS::S3::Bucket
+    DeletionPolicy: Delete
+    Properties:
+      AccessControl: Private
+      BucketName: !Ref S3Bucket
+Outputs:
+  S3BucketRoleArn:
+    Description: Spleen S3 Bucket
+    Value: !GetAtt SpleenBucket.Arn
+    Export:
+      Name: !Sub "${AWS::StackName}-${S3Bucket}"
