+++ service.yml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: spleen-service
+spec:
+  selector:
+    app: spleen
+  ports:
+  - protocol: TCP
+    port: 8082
+    targetPort: 8082
+  type: LoadBalancer
